:truck: rename components - game info -> game rules - info button -> game info
