+++ src/components/frame/with-buttons/game-info/MainInfo.tsx
@@ -1,8 +1,8 @@
 import * as styles from './index.css';
 
-import type { infoType } from '@/components/frame/with-buttons/InfoButton/types';
 import ColoredButton from '@/components/button/ColoredButton';
 import SettingTextButton from '@/components/button/SettingTextButton';
+import type { infoType } from '@/components/frame/with-buttons/game-info/types';
 
 interface mainProps {
   onClose?: () => void;
