+++ src/components/frame/with-buttons/game-info/TierInfo.tsx
@@ -1,8 +1,8 @@
 import { indexProps } from './types';
 
-import * as styles from '@/components/frame/with-buttons/InfoButton/index.css';
 import SettingNavigationButton from '@/components/button/SettingNavigationButton';
-import { LOW_TIER_INFO, HIGH_TIER_INFO } from '@/components/imagebox/constants';
+import * as styles from '@/components/frame/with-buttons/game-info/index.css';
+import { HIGH_TIER_INFO, LOW_TIER_INFO } from '@/components/imagebox/constants';
 import TierBox from '@/components/imagebox/TierIconBox';
 import { tierType } from '@/components/imagebox/types';
