+++ src/components/frame/with-buttons/game-info/index.css.ts
